refactor(readiness): split contract modules
Keep production_readiness.py as the exact public compatibility facade while extracting values, authentication, building, evidence derivation, bounded JSON, structural validation, and parser validation into cohesive modules. Split the discovered test suite into non-discovered responsibility mixins so all 45 fully qualified test IDs remain unchanged.

Ultraworked with [omo](https://github.com/code-yeongyu/oh-my-openagent)

Constraint: Preserve every public symbol, canonical JSON byte, HMAC/trust boundary, error contract, and deterministic concurrency assertion.
Rejected: SIZE_OK exceptions | production and test seams split cohesively below 250 pure LOC without changing behavior.
Confidence: high
Scope-risk: moderate
Directive: Keep readiness authentication, parsing, evidence derivation, and concurrency tests in their dedicated modules.
Tested: 45 readiness tests; 46 operations/memory tests; 7,404 full tests; canonical JSON and 38-symbol equality; strict BasedPyright; Ruff; compileall; pure-LOC gate; git diff --check.
Not-tested: Remote CI after this behavior-identical modular split.
Plan: .omo/plans/omh-domain-skill-quality-upgrades.md
